+++ Install-SRSReport/Private/Show-Progress.ps1
@@ -1,0 +1,90 @@
+#region Function Show-Progress
+Function Show-Progress {
+<#
+.SYNOPSIS
+    Displays progress info.
+.DESCRIPTION
+    Displays progress info and maximizes code reuse.
+.PARAMETER Actity
+    Specifies the progress activity. Default: 'Running Install Please Wait'.
+.PARAMETER Status
+    Specifies the progress status.
+.PARAMETER CurrentOperation
+    Specifies the current operation.
+.PARAMETER Step
+    Specifies the progress step. Default: $Global:Step ++.
+.PARAMETER ID
+    Specifies the progress bar id.
+.PARAMETER Delay
+    Specifies the progress delay in milliseconds. Default: 400.
+.PARAMETER Loop
+    Specifies if the call comes from a loop.
+.EXAMPLE
+    Show-Progress -Activity 'Running Install Please Wait' -Status 'Uploading Report' -Step ($Step++) -Delay 200
+.INPUTS
+    None.
+.OUTPUTS
+    None.
+.NOTES
+    This is an private function should tipically not be called directly.
+    Credit to Adam Bertram.
+.LINK
+    https://adamtheautomator.com/building-progress-bar-powershell-scripts/
+.LINK
+    https://SCCM.Zone/
+.LINK
+    https://SCCM.Zone/Install-SRSReport-GIT
+.LINK
+    https://SCCM.Zone/Install-SRSReport-ISSUES
+.COMPONENT
+    RS
+.FUNCTIONALITY
+    RS Catalog Item Installer
+#>
+    [CmdletBinding()]
+    Param (
+        [Parameter(Mandatory=$false,Position=0)]
+        [ValidateNotNullorEmpty()]
+        [Alias('act')]
+        [string]$Activity = 'Running Install Please Wait',
+        [Parameter(Mandatory=$true,Position=1)]
+        [ValidateNotNullorEmpty()]
+        [Alias('sta')]
+        [string]$Status,
+        [Parameter(Mandatory=$false,Position=2)]
+        [ValidateNotNullorEmpty()]
+        [Alias('cro')]
+        [string]$CurrentOperation,
+        [Parameter(Mandatory=$false,Position=3)]
+        [ValidateNotNullorEmpty()]
+        [Alias('pid')]
+        [int]$ID = 0,
+        [Parameter(Mandatory=$false,Position=4)]
+        [ValidateNotNullorEmpty()]
+        [Alias('ste')]
+        [int]$Step = $Script:Step ++,
+        [Parameter(Mandatory=$false,Position=5)]
+        [ValidateNotNullorEmpty()]
+        [Alias('del')]
+        [string]$Delay = 400,
+        [Parameter(Mandatory=$false,Position=5)]
+        [ValidateNotNullorEmpty()]
+        [Alias('lp')]
+        [switch]$Loop
+    )
+    Begin {
+        If ($Loop) { $Script:Steps ++ }
+        $PercentComplete = $($($Step / $Steps) * 100)
+    }
+    Process {
+        Try {
+            ##  Show progress
+            Write-Progress -Activity $Activity -Status $Status -CurrentOperation $CurrentOperation -ID $ID -PercentComplete $PercentComplete
+            Start-Sleep -Milliseconds $Delay
+        }
+        Catch {
+            Throw (New-Object System.Exception("Could not Show progress status [$Status]! $($_.Exception.Message)", $_.Exception))
+        }
+    }
+}
+#endregion