+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog for Install-SRSReport
 
+## 1.1.0 - 2020-03-26
+
+* Fixed default parameters, all required parameters will be requested if running the script without parameters.
+* Fixed report upload when non report files are present in the upload folder.
+* Fixed incorrect `Add-RISQLExtension` parameter set names.
+* Added `Show-Progress` function.
+* Added progress indicators.
+* Install module will always show verbose info.
+* Removed warning messages.
+* Removed default `ReportFolder` parameter value.
+* Updated description links.
+
 ## 1.0.1 - 2020-03-16
 
 * Added folder for organization